api/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==3.7.1
attrs==25.3.0
bcrypt==4.0.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cramjam==2.9.1
dataclasses-json==0.5.14
distro==1.9.0
ecdsa==0.19.1
fastapi==0.104.1
fastparquet==2024.11.0
frozenlist==1.5.0
fsspec==2024.12.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jellyfish==1.0.3
jiter==0.9.0
json_repair==0.39.1
jsonpatch==1.33
jsonpointer==3.0.0
kaleido==0.2.1
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.49
langchain-neo4j==0.4.0
langchain-openai==0.3.11
langchain-text-splitters==0.3.7
langgraph==0.3.21
langgraph-checkpoint==2.0.23
langgraph-prebuilt==0.1.7
langgraph-sdk==0.1.60
langsmith==0.3.19
marshmallow==3.26.1
multidict==6.2.0
mypy-extensions==1.0.0
narwhals==1.32.0
neo4j==5.28.1
neo4j-graphrag==1.6.1
networkx==3.4.2
numexpr==2.10.2
numpy==1.26.4
openai==1.69.0
openapi-schema-pydantic==1.2.4
orjson==3.10.16
ormsgpack==1.9.1
packaging==23.2
pandas==2.1.3
passlib==1.7.4
plotly==6.0.1
propcache==0.3.1
psycopg2-binary==2.9.9
pyarrow==19.0.1
pyasn1==0.6.1
pydantic==2.11.1
pydantic-settings==2.8.1
pydantic_core==2.33.0
PyJWT==2.8.0
PyMySQL==1.1.0
pyodbc==5.0.1
pypdf==5.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
tenacity==8.5.0
tiktoken==0.9.0
tqdm==4.67.1
types-PyYAML==6.0.12.20250326
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.24.0
xxhash==3.5.0
yarl==1.18.3
zstandard==0.23.0
langchain_community==0.3.20
langchain-google-genai==2.1.2
langchain_mistralai
langchain_anthropic
langchain_ollama
langchain_aws
websockets
language-tool-python
ijson
cryptography

